+++ truffle/com.oracle.truffle.api.debug/src/com/oracle/truffle/api/debug/Debugger.java
@@ -269,26 +269,6 @@ public final class Debugger {
     }
 
     /**
-     * Sets a breakpoint to halt at any node holding a specified <em>tag</em>.
-     * <p>
-     * If a breakpoint <em>condition</em> is applied to the breakpoint, then the condition will be
-     * assumed to be in the same language as the code location where attached.
-     *
-     * @param ignoreCount number of hits to ignore before halting
-     * @param tag
-     * @param oneShot if {@code true} breakpoint removes it self after a hit
-     * @return a new breakpoint, initially enabled
-     * @throws IOException if the breakpoint already set
-     * @since 0.9
-     */
-    @SuppressWarnings({"static-method", "deprecation"})
-    @Deprecated
-    @TruffleBoundary
-    public Breakpoint setTagBreakpoint(int ignoreCount, com.oracle.truffle.api.instrument.SyntaxTag tag, boolean oneShot) throws IOException {
-        throw new UnsupportedOperationException();
-    }
-
-    /**
      * Gets all existing breakpoints, whatever their status, in natural sorted order. Modification
      * save.
      *