+++ truffle/com.oracle.truffle.api.instrumentation/src/com/oracle/truffle/api/instrumentation/SourceSectionFilter.java
@@ -659,21 +659,18 @@ public final class SourceSectionFilter {
 
             private final SourceSection[] sourceSections;
 
-            @SuppressWarnings("deprecation")
             SourceSectionEquals(SourceSection... sourceSection) {
                 this.sourceSections = sourceSection;
                 // clear tags
                 for (int i = 0; i < sourceSection.length; i++) {
-                    sourceSections[i] = sourceSection[i].withTags();
+                    sourceSections[i] = sourceSection[i];
                 }
             }
 
             @Override
-            @SuppressWarnings("deprecation")
             boolean isIncluded(Set<Class<?>> providedTags, Node instrumentedNode, SourceSection s) {
-                SourceSection withoutTags = s.withTags();
                 for (SourceSection compareSection : sourceSections) {
-                    if (withoutTags.equals(compareSection)) {
+                    if (s.equals(compareSection)) {
                         return true;
                     }
                 }
@@ -712,12 +709,11 @@ public final class SourceSectionFilter {
 
             private final SourceSection[] sourceSections;
 
-            @SuppressWarnings("deprecation")
             RootSourceSectionEquals(SourceSection... sourceSection) {
                 this.sourceSections = sourceSection;
                 // clear tags
                 for (int i = 0; i < sourceSection.length; i++) {
-                    sourceSections[i] = sourceSection[i].withTags();
+                    sourceSections[i] = sourceSection[i];
                 }
             }
 
@@ -727,15 +723,13 @@ public final class SourceSectionFilter {
             }
 
             @Override
-            @SuppressWarnings("deprecation")
             boolean isRootIncluded(Set<Class<?>> providedTags, SourceSection rootSection) {
                 if (rootSection == null) {
                     return false;
                 }
 
-                SourceSection withoutTags = rootSection.withTags();
                 for (SourceSection compareSection : sourceSections) {
-                    if (withoutTags.equals(compareSection)) {
+                    if (rootSection.equals(compareSection)) {
                         return true;
                     }
                 }