+++ truffle/com.oracle.truffle.api/src/com/oracle/truffle/api/source/Source.java
@@ -29,18 +29,17 @@ import java.io.ByteArrayInputStream;
 import java.io.File;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.Reader;
+import java.net.URI;
+import java.net.URISyntaxException;
 import java.net.URL;
 import java.nio.charset.Charset;
 import java.nio.file.spi.FileTypeDetector;
+import java.util.Objects;
 
 import com.oracle.truffle.api.CompilerAsserts;
 import com.oracle.truffle.api.TruffleLanguage.Registration;
-import com.oracle.truffle.api.nodes.Node;
-import java.io.InputStreamReader;
-import java.util.Objects;
-import java.net.URI;
-import java.net.URISyntaxException;
 
 /**
  * Representation of a source code unit and its contents. Source instances are created by using one
@@ -63,8 +62,8 @@ import java.net.URISyntaxException;
  *
  * {@link SourceSnippets#fromURL}
  *
- * Each URL source is represented as a canonical object, indexed by the URL. Contents are
- * <em>read eagerly</em> once the {@link Builder#build()} method is called.
+ * Each URL source is represented as a canonical object, indexed by the URL. Contents are <em>read
+ * eagerly</em> once the {@link Builder#build()} method is called.
  *
  * <h3>Source from a literal text</h3>
  *
@@ -801,22 +800,11 @@ public abstract class Source {
      */
     public final SourceSection createSection(String identifier, int startLine, int startColumn, int charIndex, int length) {
         checkRange(charIndex, length);
-        return createSectionImpl(identifier, startLine, startColumn, charIndex, length, SourceSection.EMTPY_TAGS);
+        return createSectionImpl(identifier, startLine, startColumn, charIndex, length);
     }
 
-    /**
-     * @deprecated tags are now determined by {@link Node#isTaggedWith(Class)}. Use
-     *             {@link #createSection(String, int, int, int, int)} instead.
-     * @since 0.12
-     */
-    @Deprecated
-    public final SourceSection createSection(String identifier, int startLine, int startColumn, int charIndex, int length, String... tags) {
-        checkRange(charIndex, length);
-        return createSectionImpl(identifier, startLine, startColumn, charIndex, length, tags);
-    }
-
-    private SourceSection createSectionImpl(String identifier, int startLine, int startColumn, int charIndex, int length, String[] tags) {
-        return new SourceSection(null, this, identifier, startLine, startColumn, charIndex, length, tags);
+    private SourceSection createSectionImpl(String identifier, int startLine, int startColumn, int charIndex, int length) {
+        return new SourceSection(null, this, identifier, startLine, startColumn, charIndex, length);
     }
 
     /**
@@ -843,7 +831,7 @@ public abstract class Source {
             throw new IllegalArgumentException("column out of range");
         }
         final int startOffset = lineStartOffset + startColumn - 1;
-        return createSectionImpl(identifier, startLine, startColumn, startOffset, length, SourceSection.EMTPY_TAGS);
+        return createSectionImpl(identifier, startLine, startColumn, startOffset, length);
     }
 
     /**
@@ -867,20 +855,10 @@ public abstract class Source {
      * @since 0.8 or earlier
      */
     public final SourceSection createSection(String identifier, int charIndex, int length) throws IllegalArgumentException {
-        return createSection(identifier, charIndex, length, SourceSection.EMTPY_TAGS);
-    }
-
-    /**
-     * @deprecated tags are now determined by {@link Node#isTaggedWith(Class)}. Use
-     *             {@link #createSection(String, int, int)} instead.
-     * @since 0.12
-     */
-    @Deprecated
-    public final SourceSection createSection(String identifier, int charIndex, int length, String... tags) throws IllegalArgumentException {
         checkRange(charIndex, length);
         final int startLine = getLineNumber(charIndex);
         final int startColumn = charIndex - getLineStartOffset(startLine) + 1;
-        return createSectionImpl(identifier, startLine, startColumn, charIndex, length, tags);
+        return createSectionImpl(identifier, startLine, startColumn, charIndex, length);
     }
 
     void checkRange(int charIndex, int length) {
@@ -1108,7 +1086,7 @@ public abstract class Source {
          * provides {@link Source#getURI()} as a persistent identification of its location. A
          * default value for the method is deduced from the location or content, but one can change
          * it by using this method
-         * 
+         *
          * @param ownUri the URL to use instead of default one, cannot be <code>null</code>
          * @return the instance of this builder
          * @since 0.15